fix: storeMinimum UX 개선
